+++ src/llm/continuous_batch_scheduler.cpp
@@ -283,7 +283,7 @@ ContinuousBatchScheduler::scheduleNextBatch() {
         if (canAddToBatch(req.get(), total_tokens + prefill_tokens, reserved_blocks_in_batch)) {
             batch.push_back(req.get());
             total_tokens += prefill_tokens;
-            reserved_blocks_in_batch += req-> static_cast<int>(allocated_blocks.size());
+            reserved_blocks_in_batch += req->allocated_blocks.size();
             active_requests_.push_back(req);
         } else {
             // Put back in queue
@@ -294,9 +294,9 @@ ContinuousBatchScheduler::scheduleNextBatch() {
     
     // Update stats
     stats_.current_batch_size = batch.size();
-    stats_.max_batch_size_seen = std::max(stats_.max_batch_size_seen,static_cast<int>(batch.size()));
+    stats_.max_batch_size_seen = std::max(stats_.max_batch_size_seen, batch.size());
     stats_.active_requests = active_requests_.size();
-    stats_.current_queue_depth = static_cast<int>(waiting_queue_.size()) + static_cast<int>(active_requests_.size()) ;
+    stats_.current_queue_depth = waiting_queue_.size() + active_requests_.size();
     
     // Emit queue-length metric so Prometheus/Grafana can visualise scheduler
     // pressure in real time.  Called under the scheduler lock so the value is
@@ -562,7 +562,7 @@ bool ContinuousBatchScheduler::canAddToBatch(
             return true;
         }
 
-        size_t blocks_needed = request-> static_cast<int>(allocated_blocks.size());
+        size_t blocks_needed = request->allocated_blocks.size();
         if (blocks_needed == 0) {
             size_t total_tokens = request->total_prompt_tokens + request->inference_request.max_tokens;
             blocks_needed = (total_tokens + config_.block_size_tokens - 1) / config_.block_size_tokens;
@@ -630,7 +630,7 @@ void ContinuousBatchScheduler::freeKVCacheBlocks(ScheduledRequest* request) {
     kv_cache_->removeSequence(request->sequence_id);
     
     spdlog::debug("Freed {} blocks for request {} (sequence {})",
-                  request-> static_cast<int>(allocated_blocks.size()), request->request_id, 
+                  request->allocated_blocks.size(), request->request_id,
                   request->sequence_id);
     
     request->allocated_blocks.clear();
